snap: Switch main branch to core24/devel build-base
The `core24` base is available in the -edge channel for
development purposes.  At this point in time it tracks the most
recent development release of Ubuntu which at the time of this
writing is 23.10, aka. Mantic.

The purpose of making this switch for the main branch is to have
early exposure to the upcoming versions of Open vSwitch and OVN,
which may be useful both for bisecting upstream issues and for
ensuring the snap build and microovn is ready for the next
release.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>