[WIP] test: Add autotest framework for functional tests
One of the possible frameworks we may use for our testing
is the autotest framework.

This patch put together a minimal set of macros and tests
to demonstrate how it could be used with LXD as a backend
for running the contained code.

While it is a bit archaic, it is very simple and self contained,
it is also well understood by the pool of people interested in
working on OVS and OVN.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>