Makefile: drop non-funtioning .PHONY target
Before this patch a .PHONY target was added presumably to select
the default behavior when just running `make`.  Remove it as it
does not work.

By default, GNU Make will select the first target not starting
with a `.` when no target is specified.  Re-order the targets to
accomplish the same thing.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>