Makefile: rename test target, update dependency handling
Rename functional test target from `func-tests` to `check-system`.

Make snapcraft pack in verbose mode.  Snapcraft defaults to build
with LXD when `core22` is the base, so drop extranous argument.

Make `check-system` target depend on the existence of the snap
rather than the `build` target.  This way tests can be run
multiple times without unconditionally rebuilding the snap every
time.

Each line in a make target is a separate shell invocation, which
is not very effective.  Use a single shell for the entire target
by ending each command with `;` and adding a continuation `\` at
the end of the line.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>