tests/basic_cluster: Fix sharing of TEST_CONTAINERS
The committed solution for sharing TEST_CONTAINERS between file
level setup and teardown functions and the individual tests does
not actually work.

This was not caught because the individual tests do their
assertions within loops, they provided false positive results when
the loop variable was empty.

Add assertion on contents of TEST_CONTAINERS in per-test setup
function to avoid false positive results.

Export the TEST_CONTAINERS variable from the file-level setup
function.  This is in line with the bats documentation and have
confirmed it works fine even when running multiple test files
in parallel.

Add a shell fragment comment at the top so that editors/IDEs enable
bash syntax highlighting when editing the file.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>