tests: Separate test definitions from file level setup
The bats framework provide no way of composing a test suite, i.e.
there is no include directive that take test definitions into
account.

For microovn we often have a set of tests that we want to run in
multiple deployment scenarios, and as such we need some way of
running the same tests for different setups.

This patch accomplishes this by:

* Move test definitions into test_helper/bats

* Move file level setup definitions into test_helper/setup_teardown

* Add test_helper/setup_loader.bash which loads the appropriate
  setup definitions based on the `$BATS_TEST_FILENAME` variable.

* Finally create symlinks in the top-level tests/ directory which
  combines a test_helper/bats/ test definition with a
  test_helper/setup_teardown/ setup definition.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>