tests/test_helper/lxd: Load openvswitch kernel module through LXD config
To improve developer ergonomics, we ensure the `openvswitch`
kernel module is loaded by making use of the
`linux.kernel_modules` LXD instance configuration option.

Doing it this way gives one less step for anyone wanting to run
the test suite on their own system.

This makes it obvious that the `tests/cloud-init.yaml` file is
badly named, so rename it.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>