github: Dynamically generate test matrix
The microovn tests are formulated in a set of test suites.  Each
test suite typically deploys an individual namespaced environments
for installing the snap, performing configuration and running
tests.

On a powerful computer this can be utilized to have the test
runner run test suites, and individual tests within each suite in
parallel.

However, the GitHub Hosted runners are quite modest and typically
provide 2 cores and 7.5 GB RAM.

In order to parallelize jobs in such an environment, we need to
run each individual suite in a separate runner.  This is
accomplished by using a test matrix.

Dynamically generate the test matrix based on contents of tests/
directory.  This ensures all testsuites are run, and avoids any
manual maintenance on addition or removal of a testsuite.

Build the snap once and share the artifact with the system test
jobs.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>