tests/cli/ovsovn: Ensure ovn CLI commands work as expected
Ensure the `ovn-nbctl`, `ovn-sbctl` and `ovn-trace` run
successfully.

While the tests are very simple, they confirm that paths are
properly configured in the environment and that the required snapd
interface plugs are declared.

Add IPv6 version of `cli_ovsovn` test suite because these
commands will talk to remote DB servers and we need to ensure that
works for both IPv4 and IPv6.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>