tests/test_helper/setup_teardown: Drop artifical array references
The `$TEST_CONTAINERS` variable is a space separated string and
not a bash array.  It can't be a bash array because that is not
supported for variables exported into the environment.

Despite this fact the variable is referenced as if it was an
array in the setup_teardown functions.  This was mainly done to
silence `shellcheck` SC2086 info-level messages, and in retrospect
this was a mistake and just confusing.

Remove incorrect array references to `$TEST_CONTAINERS` variable.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>