tests: Add cluster scale-up test
Add a test case that incrementally scales up a cluster from 1 to
4 nodes, (re-)running the cluster tests after each scaling event.

To work around the fact that bats-core does currently not provide
any way of programmatically composing tests, we use the workaround
provided in bats-core/bats-core#241.  Essentially we use a bats
test definition file as a wrapper to invoke tests from a different
bats test definition file.

Note that the outer bats runner will perform validation on the
number of tests ran based on TAP output, so it is important that
the inner bats runner uses a different format for its output.

We should monitor the progress of bats-core/bats-core#349 and move
to something like that when it materializes.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>