ovn/refresh: Don't restart ovn-controller on cluster change
Before this commit the `ovn-controller` process would be restarted
whenever a node is added or removed from the cluster.

This is not necessary as the `ovn-controller` will read its
remote DB configuration from the local Open vSwitch instance
(`external_ids:ovn-remote`).

In the event of change of certificates, the `ovn-controller` will
also re-read on-disk certificate data without the need for a
restart.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>