snap: Build proper version string for snap
The main component of MicroOVN is OVN itself, thus we build the
version string around the upstream version of OVN.

The snap consumes the OVN binaries from the deb in the Ubuntu
version corresponding to the build base, benefitting from the
security and other support going into that artifact.  While it
could be useful to know the distro version (i.e. `0ubuntu0.22.04.1`,
we deliberately exclude it from the end user visible snap version
string because snaps are universal packages that could be
installed on any distribution.  And having ubuntu in the version
string would be confusing for those users.

In addition to the upstream OVN version, we include an abbreviated
commit hash from the `microovn` repository.  These two pieces
together allows exact identification of what version of the snap is
being consumed for any reported issues from the userbase.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>