snap: Add man pages.
The upstream OVS and OVN projeccts provide detailed documentation
in their manpages.  It would be a shame to deprieve the user of
this documentation.

There have been efforts to bring man page support to the snap
infrastructure [0][1], but it has unfortunately not come to
fruition.  We need to bring the required components ourself.

Example to view the OVN Architecture documentation:

    microovn.man ovn-architecture

0: https://forum.snapcraft.io/t/support-for-man-pages/2299
1: snapcore/snapd#8079
Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>