Use systemd journal for logging.
At present the OVS and OVN services provided by MicroOVN logs to
file.

This is a problem, because there is no mechanism for rotating
these logfiles, and as a consequence they will grow without
bounds potentially depleting all available storage.

The best practice for snaps is to log to the systemd journal [0].

Note that due to how the `ovs-ctl` and `ovn-ctl` scripts work,
there is no way to avoid the log files from being created.  It
may also be useful to have log files configured for live
debugging sessions.

0: https://forum.snapcraft.io/t/best-practice-for-logging/1210

Closes-Bug: #2039582
Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>