CI: Install snapcraft from latest/edge on main branch.
The main branch tracks the current Ubuntu development release,
this gives us early warning on coming issues for future releases
which we can use both to prepare and to feed issues to projects
we depend on early.

We now have a concrete example where builds are blocked and a fix
is available on the snapcraft latest/edge channel.

Use the latest/edge channel for snapcraft on main branch, with
fallback to the latest/stable channel on other branches.

The fallback minimizes the manual work required when branching for
future stable releases.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>