commands: Refactor and export OvsdbSpec struct
Refactor: Additional properties were added to this struct that
are related to, and can be defined for, any OVN/OVS database.
Also, property 'Target' was renamed to 'SocketPath' to better
match its purpose.

Export: This structure is usefull in other modules which was
prviously achieved via exported helper functions. Helper functions
are now removed and the structure is exported directly.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>