api: Endpoint for expected OVSDB schema version
This change adds API endpoint, and an associated client function, for
determing expected schema version of specific OVS/OVN database.

Expected schema version is the one that was supplied with the OVN/OVS
package for given database. It may differ from the schema version used
by currently running OVN/OVS processes. Difference between the running
and the expected schema versions is normal situation during the
MicroOVN cluster upgrade, when an individual MicroOVN node may have
been upgraded, but the cluster-wide schema upgrade for the database
was not yet triggered.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>