ovsdb: Manually trigger OVSDB cluster schema upgrade
This change takes over the responsibility for triggering OVN OVSDB
cluster schema upgrades from the "ovn-ctl" script, and lets the
MicroOVN to trigger the schema upgrade when all the nodes are ready.

OVN NB and SB services are now started with
"--db-cluster-schema-upgrade=no" which lets the "ovn-ctl" script know
that it should not attempt to trigger schema upgrade on start of the
service.

Instead, microovn daemon checks the necessity of schema upgrade on its
start. If the schema upgrade for NB or SB database is necessary, it
will coordinate with the rest of the cluster members, and trigger the
schema upgrade when all the nodes are ready for it.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>