snap: Build docs python dependencies from source.
The docs part populates a virtualenv with PIP for the tooling
required to build the docs.  PIP will by default download binary
wheels from PyPI when available, however not all architectures
have pre-built wheels available.

This is problematic because at present a build may pass in the
gate only to fail further down in the post-merge pipeline.

To mitigate this we change the docs Makefile to instruct PIP to
avoid using binary wheels so that any missing build dependencies
can be detected in the gate.

Related issue: canonical/sphinx-docs-starter-pack#238

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>