tests: Measure data path uptime during upgrade.
Create a network namespace to simulate a container/virtual
machine workload running on a node without the central service.

Measurements are performed by starting a ping in the background
inside the workload before performing certain steps, and then
reaping the ping process, reading the resulting statistics.

The target for the ICMP echo requests is a gateway router which
the OVN Controller installs into the local Open vSwitch instance,
and we have confirmed that this is sufficient to detect side
effect of an sincorrect upgrade sequence.

Results are checked at two points in the upgrade process:
1. After upgrading central nodes.  This measurement ensures that
   MicroOVN handles the upgrade in such a way so that a central
   first upgrade does not have datapath impact.  Zero packet loss
   is expected.

2. After upgrading chassis nodes with workloads.  This measurment
   ensures that MicroOVN handles service life cycle on snap
   refresh with minimal downtime for workloads.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>