API: Endpoint for fetching active OVSDB version
This change adds MicroOVN API endpoint and associated client function
that allows fetching active OVSDB schema version for given datbase.

This API is available on any node. If the node that receives request
about OVN Northbound or Southbound DBs does not run "central"
services, the request will be forwarded to a node that does.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>