tests: Determine initial version from filename
Upgrade tests need to determine the initial MicroOVN version so that
it can be deployed, and then upgraded to the MicroOVN built from
source. This verifies that the upgrade path is possible.

This initial version was previously controlled by env variable
`MICROOVN_SNAP_CHANNEL`, but the limitation was that in a single run,
we could define only one set of upgrade tests (from
`MICROOVN_SNAP_CHANNEL` to currently built).

This change introduces a mechanism to determine the initial MicroOVN
version based on the test filename instead. With this approach, we can
create multiple symbolic links (e.g. `upgrade_22.03.bats ->
upgrade.bast`, `upgrade_24.03.bats -> upgrade.bats`) to trigger
multiple parallel upgrade tests.

Upgrade tests with the original/generic name (e.g. `upgrade.bats`)
will upgrade from the version that's in the default channel for
the MicroOVN snap.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>