snap: Run golangci-lint as part of build.
There is currently no lint nor static checks for the MicroOVN
Golang code.

Surveying some of our neighbouring projects, the most widely used
tool to do this appears to be golangci-lint.

Successfully running the tool requires build dependencies to be
installed and build environment variables properly set up.

Developer ergonomics are important to us, and we want the same
experience in local development environemnts as in our gate.

Running the tool as part of the build gives us consistent results
in both environments and relieves the developer of the burden of
manually installing build dependencies to perform the checks
locally.

We will gradually enable the default linters [0] one by one making
the update to `microovn/.golangci.yml` that goes along with any
code changes document which check mandated the code changes.

0: https://golangci-lint.run/usage/linters/

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>