Better (enabling|disabling) of central services
This changes the central service enabling to, regenerate the env file
before enabling, generate certificates to allow secure communication
with the other nodes.

This changes the central service disabling to properly leave the cluster
and transfer ownership if they are the leader.

These changes are to be followed by a change involving an api endpoint
for environment refresh.

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>