regenerate_ca: Fix requests not being forwarded
API endpoint for regenerating CA certificates was broken after
recent upgrade to `microcluster` library[0] that changed signature
of the `State.Cluster()` method. This method now accepts boolean
value that must be set to `true` if we want recipient to properly
recognized the request with the `client.IsNotification()` method.

Current code was not properly recognizing the initial request from the
client and as a consequence, it did not trigger re-issue of the CA
certificate and subsequent notification to the rest of the cluster.

[0] https://github.com/canonical/microovn/commit/5b42b2ed7a291de6c5a5551b5b3edfbb35a45adb

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>