Removal of deprecated central snap service
This removes the microovn.central service. This has been coming a long
time and removing it should ease confusion between the central service
and the central snap service.

This service has been discouraged for a long time and doesn't handle its
"child" services particularly well and people shouldn't really be
interacting with these snap services anyway as we have enablement and
disablement of services within the codebase

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>