tests: Set upgrade channel to 24.03/beta
This branch of MicroOVN is based on OVN 24.09. That means that
it can no longer directly upgrade from current default MicroOVN
snap channel (22.03).

These changes set the base channel for upgrade tests to 24.03/beta

Note: Since the MicroOVN 24.03 does not have a 'stable' channel yet,
we need to use 'beta'

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>