Add defaults for initial and and connect addresses
Previously if the enviroment file was generated before any central
services had been enabled, then the initial and connect fields would be
empty, stopping northd from connecting to south and north databases.

This fix gives these defaults, which will only be used in the case of a
no central being found, in which case a central will always be generated
on the local machine, so the default addresses will be accurate.

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>