snap: Skip golangci-lint on riscv64.
The golangci-lint snap is unfortunately not built for riscv64.

The snapcraft advanced grammar [0] does unfortunately not allow
us to express different build dependencies on different
architectures without repetition of the list of dependencies.

Making use of composition at the YAML level is also not possible
because merging sequences is just not supported [1].

0: https://snapcraft.io/docs/snapcraft-advanced-grammar
1: https://github.com/yaml/yaml/issues/35#issuecomment-496184265
Reported-at: alexmurray/golangci-lint-snap#3
Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: Theodoros Vakalopoulos <theovakalopoulos@gmail.com>