bgp: Learn default route from the network.
As part of our mission to provide an opinionated routed data
center solution that require minimal to no configuration,
automatically learning the default route is an important part.

In CLOS topologies that make use of eBGP as routing protocol,
similar to what is described in RFC 7938, Tier n devices typically
only learn a default route.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>