cert-renew: Don't refresh services that are not present
Automatic script that renews certificates near the expiration
was attempting to check/renew certificates for all services, regardless
of whether they were present on the node or not.
This did not affect renewal of existing certificates, but it did log errors,
and the systemctl service run exited with non-zero exit code.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>