Only reissues certs when CA is actually changed
Certificates will only be reisssued to all nodes when the CA private key
or certificate is actually changed, this means when we run set-ca with
the same certificate multiple times we don't reissue all of the
certificates which saves lots of compute time.

This is related to ongoing microovn charm work, specifically the
tls-certificates integration.

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>