CI: Trigger full test suite only by push to the 'main' branch
When CI run is triggered by a 'push' action, it should run tests only
if the push was made to the `main` branch. This avoids duplicate CI runs
when a PR is opened from a branch of the primary repository instead of a fork.

This does not affect CI runs triggered by pull requests.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>