config: Allow configuring "ovn.central-ips" config option via CLI.
This change contains a new feature that allows user-specified configuration to be stored
in the MicroOVN database. This provides the user with a convenient way to set cluster-wide
configuration. MicroOVN will store the value in its distributed database and perform the
necessary steps to apply the change on each cluster member.

The first implemented user option is "ovn.central-ips" for specifying IP addresses of OVN central
nodes to which the controller (and clients) should connect to. It overrides the default behavior
in which IP addresses of MicroOVN nodes with "central" service enabled are used.
This value is expected in the format of comma-separated IP addresses.

Example:
microovn config set ovn.central-ips 10.0.0.1,10.0.0.2,10.0.0.3

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>
(cherry picked from commit d2673f054dda14eb9f02bc796b3f7031e4f72313)