join: Change default services enabled on join.
Previously, the "join" hook enabled 'central' service on the node if there were
less than three central services enabled on the cluster. With the addition of
an option to manually configure external OVN Central cluster, the join hook will no
longer enable the central service if this option is set.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>