node: Fix incomplete remote addresses in northd
MicroOVN used incomplete list of NB/SB cluster members when
starting the ovn-northd service. When central service was enabled,
local ovn.env file, containing list of IP addresses of NB/SB
cluster members, was not updated before starting of the northd
service. This caused northd service, on a node that just enabled
the central service (or joined the cluster), to miss its own IP address.

This issue was partially hidden by the fact that all other northd
services in the cluster had complete list of remote IPs. It was
only the last joining member that was incomplete.
For the problem to manifest fully, the northd instance with the
incomplete list of remote addresses would have to aquire lock, and
the NB/SB database that was missing from its list would have to
become the cluster leader.

A new test was added to ensure that all northd instances in the
cluster have complete list of IPs of the NB/SB cluster members.

As a result of this operation reordering (ovn.env is regenerated before
enabling NB/SB/northd services), we also need to adjust the way we
generate the OVN_INITIAL_{NB,SB} variables. Previously, this variable
contained an IP address of a "first" microovn node that enabled "central"
service. This allowed the initial node to initialize clustered NB/SB
databases, and subsequently added nodes to join through the initial node.
This worked fine before service control. However now that we are free to
disable and re-enable services on any node it poses a problem in a scenario
where we disable/re-enable central service on the "first" node. When the
central service is re-enabled on the "first" node, the OVN_INITIAL_* IPs would
point to the node itself, causing it to initialize a new cluster instead
of joining the existing cluster that runs on other nodes.

The new approach is to put "local node's IP" into the OVN_INITIAL_* variables
only if we have exactly one cluster member with "central" service enabled. This
allows the first node to initialize clustered databases. As the cluster grows,
on each node we put into this variable an IP address of "other" node that already
runs central service, which allows them to join an existing cluster.

Examle:
(First node is bootstrapped)

node-1:
  * IP: 10.0.0.1
  * OVN_INITIAL_*: 10.0.0.1

(second node is added)

node-2:
  * IP: 10.0.0.2
  * OVN_INITIAL_*: 10.0.0.1

node-1:
  * IP: 10.0.0.1
  * OVN_INITIAL_*: 10.0.0.2

Fixes: ad657d813db24fad950171ddd454292f7b9d5966
Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>