node: Fix incomplete remote addresses in northd
MicroOVN used incomplete list of NB/SB cluster members when
starting the ovn-northd service. When central service was enabled,
local ovn.env file, containing list of IP addresses of NB/SB
cluster members, was not updated before starting of the northd
service. This caused northd service, on a node that just enabled
the central service (or joined the cluster), to miss its own IP address.

This issue was partially hidden by the fact that all other northd
services in the cluster had complete list of remote IPs. It was
only the last joining member that was incomplete.
For the problem to manifest fully, the northd instance with the
incomplete list of remote addresses would have to aquire lock, and
the NB/SB database that was missing from its list would have to
become the cluster leader.

A new test was added to ensure that all northd instances in the
cluster have complete list of IPs of the NB/SB cluster members.

Fixes: ad657d813db24fad950171ddd454292f7b9d5966
Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>