tests: Run tests from pre-built LXD image.
Current tests use clean ubuntu:lts LXD images in which
they usually install microovn Snap during the test setup phase.
This is a costly process and adds a significant amount of run time
to the tests.
We can improve this by pre-building an LXD image (based on ubuntu:lts)
with the microovn snap already installed, which is then used to spin
up containers by the tests. On a reasonably performant desktop PC, this
improves test run time by roughly 50% for basic_cluster.bats which uses
3 containers. The time savings are even better for tests that use more
containers.

This approach is a suitable replacement for most of the system tests.
However, there are few that will keep the old workflow:
* lifecycle - The MicroOVN snap needs to be installed as part of the test
              and therefore can't be pre-installed on the image.
* ovsdb_schema_upgrade - This upgrade test needs to start with MicroOVN
                         installed from the snap store.
* tls_cluster - This test uses VMs. It's debatable whether it's worth it
                to pre-build an LXD VM image just for this one test suite.
* upgrade - This upgrade test needs to start with MicroOVN installed from the
            snap store.

As a consequence of using custom LXD image for testing, we no longer
support specifying custom image through MICROOVN_TEST_CONTAINER_IMAGE
env variable.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>