bgp: Enable ECMP, use BFD when peer has it configured.
When multiple links are configured, the default behavior of the
routing protocol suite is to select and redistribute one route.

We want multiple paths with equal cost to be redistributed to
allow load sharing across the links (ECMP).

When load sharing across links, it is imperative to detect and act
on failure immediately, to avoid every other packet being sent into
a void.

Bidirectional Forwarding Detection (BFD) is the way to do this, and
BIRD supports a passive mode where it only enables this after
receiving BFD packets from its peer, allowing the user to opt into
the feature through configuration on the Top-of-Rack (ToR) switch.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>