docs: Explanation and How-To for BGP
This commit adds documentation for the BGP integration feature of
MicroOVN.

The Explanation goes a little bit under the hood, giving context on
how the feature works upstream and what steps MicroOVN performs when
the BGP integration is enabled.

The How-To guide provides simple steps for enabling and disabling the
integration.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>