config: Add validation for config option input
The interface for setting config options (microovn config set) is missing validation on user input.
This can cause silent failures and other runtime errors.

Validation has been added to the server-side API for config options.
Each config option should have a linked validator (pointer to a function that returns properly
formatted errors based on input data) to check the input and inform the user of the failed
validation and config update.

A testcase was also added to check that a call to microovn config set ovn.central-ips with
unparseable values fails as expected.

Signed-off-by: Matteo Perin <matteo.perin@canonical.com>