bgp: Return error message when enabling BGP without VRF module
The control plane command to enable the BGP feature fails silently
when the VRF given as configuration cannot be created due to having
no vrf module loaded in the kernel.
This can lead to confusion and unexpected behaviour at runtime.

A quick check has been, therefore, added before trying to enable
the feature. It works by looking for a reference to a running vrf module in /sys/module.

Given that this (and other) preemptive checks can lead to a failure
before the netplan config yaml is created, a check that this file
actually exists has been also added during the cleanup phase.

A testcase has also been added to check the CLI command returns the
expected error message, given the above conditions.

Signed-off-by: Matteo Perin <matteo.perin@canonical.com>