snap/docs: Bump rust version for Python dependencies.
The snap builds the documentation so that we can ship a plain text
rendering of our documentation with the snap.

There are direct and indirect dependencies required for the Sphinx
documentation framework, some of which require a rust compiler.

Note that since we do not ship the binary artifacts of the
documentation Python dependencies, we allow the use of pre-built
wheels from PyPI.  This issue becomes evident when building on
architectures that do not have pre-built wheels.

Rust is still developing at a neck breaking pace and the original
compiler shipped with Noble is now too old.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>
(cherry picked from commit 3e51f75cf1e38f98cb54e5bb6f00ba38455e0afb)